Fix resolution callback running twice
When the executing script isn't known, a faux dependency record is created, which will then be absorbed when the next dependency finishes. Causing the faux dependency to init can roll right through to finalization, which will run the resolution callbacks, then when the faux dependency is absorbed, the callback is run again. To keep this from happening, the faux dependency shouldn't be initialized.